pyproject.toml
[project]
name = "deepagents-project"
version = "0.1.0"
description = ""
authors = [
    {name = "joaopjp11",email = "joaopires2000@hotmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "deepagents (>=0.1.4,<0.2.0)",
    "langchain-google-genai (>=3.0.0,<4.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "fastapi (>=0.120.0,<0.121.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "pydantic (>=2.12.3,<3.0.0)",
    "pandas (<2.3.0)",
    "llama-index (>=0.14.6,<0.15.0)",
    "chromadb (>=1.3.0,<2.0.0)",
    "llama-index-vector-stores-chroma (>=0.5.3,<0.6.0)",
    "llama-index-llms-google-genai (>=0.6.2,<0.7.0)",
    "llama-index-embeddings-google-genai (>=0.3.1,<0.4.0)",
    "llama-index-embeddings-huggingface (>=0.6.0,<0.7.0)",
    "datasets (>=4.3.0,<5.0.0)",
    "scikit-learn (>=1.7.2,<2.0.0)",
    "sentence-transformers (>=5.1.2,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [{include = "*", from = "src"}]
